fix(pr6-followup): cast ff_team_id to long, not int (Spark ANSI CAST_OVERFLOW) (#209)
PR #208's DEFCON re-ingest (run_id=240420217744914) failed with:

  [CAST_OVERFLOW] The value 'long' of the type "LONG" cannot be cast to
  "INTEGER" due to an overflow.

despite all team_id values fitting in INT range. Root cause: Spark
ANSI mode rejects LONG-to-INT casts STATICALLY (based on type
analysis, not actual values). The expression

  F.when(F.col("ff_teammate"), F.col("act_team_id"))
  .otherwise(F.expr("element_at(array_except(...), 1)"))

was inferred as LONG (because array_except / element_at returns LONG
by default for non-trivially-typed array operands), and the trailing
.cast("int") then failed ANSI's static check.

Fix: change .cast("int") to .cast("long"). This matches the
credits_schema StructField for defender_team_id (LongType), so it's
the right target type anyway. The pre-PR-6-followup hardcoded
F.lit(0).cast("int") worked only because lit(0) is a static INT,
giving Spark no opportunity to widen.

Pre-push verified: ruff clean; pytest test_trigger_dbt_job 9/9 pass.

+++ src/ingestion/defcon_lite_360.py
@@ -280,7 +280,13 @@ def process_360_matches(
                 # team_id from the per-match list, take the remaining element.
                 F.expr("element_at(array_except(_match_team_ids, array(act_team_id)), 1)")
             )
-            .cast("int"),
+            # cast to long (not int): credits_schema declares
+            # defender_team_id as LongType, AND Spark ANSI mode rejects
+            # LONG-to-INT casts statically (the array_except / element_at
+            # combinator returns LONG by default — observed 2026-04-27 on
+            # PR-#208 deploy: [CAST_OVERFLOW] thrown despite all team_id
+            # values fitting in INT range).
+            .cast("long"),
         )
         .drop("_match_team_ids")
     )
